+++ server/application-server/src/main/java/de/tum/in/www1/hephaestus/gitprovider/team/TeamRepository.java
@@ -1,15 +1,12 @@
 package de.tum.in.www1.hephaestus.gitprovider.team;
 
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface TeamRepository extends JpaRepository<Team, Long> {
     Team findFirstByName(String teamName);
 
-    List<Team> findTeamByName(String teamName);
+    List<Team> findAllByName(String name);
 }
-